==> 6) Individual experiments/2022-06-21 - Illumina paired-end 2sg - correct reference/01_R_scripts/04_look_up_sgRNAs_run1.R <==
## 2022-06-21


# Load packages and source code -------------------------------------------

CRISPR_root_directory <- "~/CRISPR_4sgRNA"
experiments_directory <- file.path(CRISPR_root_directory, "6) Individual experiments")
first_nanopore_dir <- file.path(experiments_directory, "2022-01-05 - first nanopore sequencing run")
source(file.path(first_nanopore_dir, "01_R_scripts", "1_R_functions", "05_looking_up_aligned_sgRNAs.R"))



# Define folder paths -----------------------------------------------------

project_dir <- file.path(experiments_directory, "2022-06-21 - Illumina paired-end 2sg - correct reference")
rdata_dir   <- file.path(project_dir, "03_R_objects")

first_illumina_trial_dir <- file.path(experiments_directory, "2022-04-21 - Illumina paired-end 2sg - first trial")
run1_rdata_path <- file.path(first_illumina_trial_dir, "03_R_objects", "01_read_in_data_run1.RData")



# Load data ---------------------------------------------------------------

load(file.path(rdata_dir, "02_annotate_dJR072_CRISPRoff_library.RData"))
load(run1_rdata_path)



# Prepare data frames for input to the CheckGuides function ---------------------------------------

extracted_df <- run1_fastq_df[, paste0("Sequence_sg", 1:2)]
names(extracted_df) <- paste0("Aligned_read_sg", 1:2)

sg_sequences_df <- CRISPRoff_df[, c("protospacer_A", "protospacer_B")]
names(sg_sequences_df) <- paste0("Sequence_sg", 1:2)
sg_sequences_df[, "Sequence_sg1"] <- substr(sg_sequences_df[, "Sequence_sg1"], 2, 20)
sg_sequences_df[, "Sequence_sg2"] <- substr(sg_sequences_df[, "Sequence_sg2"], 2, 20)

rm(run1_fastq_df)



# Look up aligned sgRNAs --------------------------------------------------

extract_df_list <- lapply(1:2, function(x) {
  CheckGuides(extracted_df, x, large_chunk_size = 500000, small_chunk_size = 50000)
})

load(run1_rdata_path)

run1_matched_df <- do.call(data.frame, c(list(run1_fastq_df[, !(names(run1_fastq_df) %in% c("Quality_sg1", "Quality_sg2"))]),
                                         extract_df_list,
                                         stringsAsFactors = FALSE
                                         )
                           )



# Save data ---------------------------------------------------------------

save(run1_matched_df,
     file = file.path(rdata_dir, "04_look_up_sgRNAs_run1.RData")
     )
